refactor:rename schema; devide into files; remove cypress to devdependencies; add message if result not found
