+++ package.json
@@ -13,7 +13,6 @@
 		"prettier": "prettier \"src/**/*.{ts,tsx}\" . --write --cache"
 	},
 	"dependencies": {
-		"@cypress/react": "^8.0.2",
 		"@firebase/util": "^1.9.7",
 		"@hookform/resolvers": "^3.9.0",
 		"@react-input/mask": "^1.2.5",
@@ -26,7 +25,6 @@
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
 		"react-hook-form": "^7.52.0",
-		"react-phone-input-2": "^2.15.1",
 		"react-phone-number-input": "^3.4.4",
 		"react-redux": "^9.1.2",
 		"react-router-dom": "^6.24.1",
@@ -37,6 +35,7 @@
 		"yup": "^1.4.0"
 	},
 	"devDependencies": {
+		"@cypress/react": "^8.0.2",
 		"@babel/core": "^7.22.11",
 		"@babel/plugin-proposal-export-default-from": "^7.22.5",
 		"@babel/plugin-proposal-throw-expressions": "^7.22.5",
