+++ src/validation/loginSchema.ts
@@ -1,6 +1,6 @@
 import * as yup from 'yup';
 
-export const LoginSchema = yup.object().shape({
+export const loginSchema = yup.object().shape({
 	phoneOrEmail: yup.string().required('Email / Phone is required'),
 	password: yup
 		.string()
